backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.*
django-axes==6.*
django-environ==0.11.*
django-cors-headers==4.*
psycopg2-binary==2.9.*
